Start organizing view setup for user signup form view model
